Implemented the class Utils which contains readSourceFile and getSourceMetaData methods.
